refactor: move magic link types to models folder
