+++ passageidentity/models/__init__.py
@@ -19,3 +19,5 @@ from passageidentity.models.update_magic_link_auth_method import (
 )
 from passageidentity.models.update_passkey_auth_method import UpdatePasskeysAuthMethod
 from passageidentity.models.update_otp_auth_method import UpdateOtpAuthMethod
+from passageidentity.models.magic_link_args import MagicLinkArgs
+from passageidentity.models.magic_link_options import MagicLinkOptions
